Reorganize general layout
This will change the layout of the page and fix the skeleton so its easier to understand. Changes the background color and adds text and underlines to key titles on the page. Also adds a new totals section skeleton where i can place the output of the future JS functions aka totals of money.